Renaming components to use PascalCase
