+++ client/components/Navbar.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { withRouter, Link } from 'react-router-dom';
+import { logout } from '../store';
+
+const Navbar = (props) => {
+  const { children, handleClick, isLoggedIn } = props;
+
+  return <div />;
+};
+
+/**
+ * CONTAINER
+ */
+const mapState = (state) => {};
+
+const mapDispatch = (dispatch) => {};
+
+// The `withRouter` wrapper makes sure that updates are not blocked
+// when the url changes
+export default withRouter(connect(mapState, mapDispatch)(Navbar));
