+++ client/index.js
@@ -1,9 +1,9 @@
 import './index.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {Provider} from 'react-redux';
+import { Provider } from 'react-redux';
 import store from './store';
-import Routes from './routes';
+import Routes from './Routes';
 
 // establishes socket connection
 import './socket';
@@ -12,5 +12,5 @@ ReactDOM.render(
   <Provider store={store}>
     <Routes />
   </Provider>,
-  document.getElementById('app')
+  document.getElementById('app'),
 );
