+++ package.json
@@ -38,6 +38,7 @@
     "redux": "^3.6.0",
     "redux-logger": "^2.8.1",
     "redux-thunk": "^2.2.0",
+    "semantic-ui-react": "^0.73.0",
     "sequelize": "^4.3.1",
     "socket.io": "^2.0.3"
   },
